mining-proxy lib
