+++ roles/mining-proxy/Cargo.toml
@@ -7,6 +7,10 @@ description = "SV2 mining proxy role"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/stratum-mining/stratum"
 
+[lib]
+name = "mining_proxy_sv2"
+path = "src/lib/mod.rs"
+
 [dependencies]
 stratum-common = { path = "../../common" }
 async-channel = "1.8.0"
